+++ app/scripts/contentscript.js
@@ -115,8 +115,8 @@ function logStreamDisconnectWarning (remoteLabel, err) {
  * @returns {boolean} {@code true} if Web3 should be injected
  */
 function shouldInjectWeb3 () {
-  return doctypeCheck() && suffixCheck()
-    && documentElementCheck() && !blacklistedDomainCheck()
+  return doctypeCheck() && suffixCheck() &&
+    documentElementCheck() && !blacklistedDomainCheck()
 }
 
 /**
@@ -176,6 +176,7 @@ function blacklistedDomainCheck () {
     'webbyawards.com',
     'cdn.shopify.com/s/javascripts/tricorder/xtld-read-only-frame.html',
     'adyen.com',
+    'gravityforms.com',
   ]
   var currentUrl = window.location.href
   var currentRegex