+++ ui/app/components/page-container/page-container-footer/page-container-footer.component.js
@@ -1,0 +1,54 @@
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import Button from '../../button'
+
+export default class PageContainerFooter extends Component {
+
+  static propTypes = {
+    onCancel: PropTypes.func,
+    cancelText: PropTypes.string,
+    onSubmit: PropTypes.func,
+    submitText: PropTypes.string,
+    disabled: PropTypes.bool,
+  }
+
+  static contextTypes = {
+    t: PropTypes.func,
+  }
+
+  render () {
+    const {
+      onCancel,
+      cancelText,
+      onSubmit,
+      submitText,
+      disabled,
+    } = this.props
+
+    return (
+      <div className="page-container__footer">
+
+        <Button
+          type="default"
+          large={true}
+          className="page-container__footer-button"
+          onClick={() => onCancel()}
+        >
+          { cancelText || this.context.t('cancel') }
+        </Button>
+
+        <Button
+          type="primary"
+          large={true}
+          className="page-container__footer-button"
+          disabled={disabled}
+          onClick={e => onSubmit(e)}
+        >
+          { submitText || this.context.t('next') }
+        </Button>
+
+      </div>
+    )
+  }
+
+}