+++ ui/app/components/selected-account/index.js
@@ -1,0 +1,2 @@
+import SelectedAccount from './selected-account.container'
+module.exports = SelectedAccount